move delivery, recorder, tracer from structures to task modules
