+++ mmrazor/models/task_modules/__init__.py
@@ -1,4 +1,7 @@
 # Copyright (c) OpenMMLab. All rights reserved.
+from .delivery import *  # noqa: F401,F403
 from .estimators import ResourceEstimator
+from .recorder import *  # noqa: F401,F403
+from .tracer import *  # noqa: F401,F403
 
 __all__ = ['ResourceEstimator']
