+++ mmrazor/structures/__init__.py
@@ -1,5 +1,2 @@
 # Copyright (c) OpenMMLab. All rights reserved.
-from .delivery import *  # noqa: F401,F403
-from .recorder import *  # noqa: F401,F403
 from .subnet import *  # noqa: F401,F403
-from .tracer import *  # noqa: F401,F403
