+++ tests/test_core/test_delivers/test_deliver_manager.py
@@ -3,7 +3,7 @@ from unittest import TestCase
 
 from mmengine import ConfigDict
 
-from mmrazor.structures import DistillDeliveryManager
+from mmrazor.models.task_modules import DistillDeliveryManager
 
 
 class TestDeliverManager(TestCase):
