+++ tests/test_core/test_delivers/test_function_outputs_deliver.py
@@ -1,7 +1,7 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 from unittest import TestCase
 
-from mmrazor.structures import FunctionOutputsDelivery
+from mmrazor.models.task_modules import FunctionOutputsDelivery
 
 
 class TestFuncOutputsDeliver(TestCase):
