+++ tests/test_core/test_recorders/test_func_outputs_recorder.py
@@ -1,7 +1,7 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 from unittest import TestCase
 
-from mmrazor.structures import FunctionOutputsRecorder
+from mmrazor.models.task_modules import FunctionOutputsRecorder
 
 
 class TestFuncOutputsRecorder(TestCase):
