+++ tests/test_core/test_recorders/test_param_recorder.py
@@ -4,7 +4,7 @@ from unittest import TestCase
 import torch
 from torch import nn
 
-from mmrazor.structures import ParameterRecorder
+from mmrazor.models.task_modules import ParameterRecorder
 
 
 class ToyModel(nn.Module):
