+++ tests/test_core/test_recorders/test_recorder_manager.py
@@ -6,7 +6,7 @@ from mmengine import ConfigDict
 from torch import nn
 from toy_mod import Toy
 
-from mmrazor.structures import RecorderManager
+from mmrazor.models.task_modules import RecorderManager
 
 
 class ToyModel(nn.Module):
